requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
groq>=0.9.0

# ─── Environment / Config ─────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── Data & Visualisation ─────────────────────────────────────────────────────
pandas>=2.2.0

# ─── Optional: PDF export ─────────────────────────────────────────────────────
# weasyprint>=62.0     # Uncomment to enable PDF generation
# Pillow>=10.0.0       # Required by weasyprint on some platforms
